refactor networks
