+++ crates/settings/src/yaml/orderbook/mod.rs
@@ -1,7 +1,6 @@
 // pub mod accounts;
 // pub mod deployer;
 // pub mod metaboards;
-pub mod network;
 // pub mod orderbook_entry;
 // pub mod sentry;
 // pub mod subgraphs;
@@ -12,14 +11,9 @@ use crate::Network;
 // use accounts::AccountsYaml;
 // use deployer::DeployerYaml;
 // use metaboards::MetaboardsYaml;
-use network::NetworkYaml;
 // use orderbook_entry::OrderbookEntryYaml;
 // use sentry::SentryYaml;
-use serde::{Deserialize, Serialize};
-use std::{
-    collections::HashMap,
-    sync::{Arc, RwLock},
-};
+use std::sync::{Arc, RwLock};
 use strict_yaml_rust::StrictYamlEmitter;
 // use subgraphs::SubgraphsYaml;
 // use token::TokenYaml;
@@ -39,7 +33,7 @@ impl OrderbookYaml {
         let document = Arc::new(RwLock::new(doc));
 
         if validate {
-            NetworkYaml::validate(document.clone())?;
+            Network::parse_networks_from_yaml(document.clone())?;
             //     SubgraphsYaml::try_from_string(&source)?;
             //     MetaboardsYaml::try_from_string(&source)?;
             //     OrderbookEntryYaml::try_from_string(&source)?;
@@ -60,16 +54,15 @@ impl OrderbookYaml {
     }
 
     pub fn get_network_keys(&self) -> Result<Vec<String>, YamlError> {
-        let networks = NetworkYaml::validate(self.document.clone())?;
+        let networks = Network::parse_networks_from_yaml(self.document.clone())?;
         Ok(networks.keys().cloned().collect())
     }
     pub fn get_network<'a>(&'a self, key: &str) -> Result<Network, YamlError> {
-        let networks = NetworkYaml::validate(self.document.clone())?;
-        let network_yaml = networks
+        let networks = Network::parse_networks_from_yaml(self.document.clone())?;
+        let network = networks
             .get(key)
             .ok_or(YamlError::KeyNotFound(key.to_string()))?;
-        let network = network_yaml.clone().try_into_network(key)?;
-        Ok(network)
+        Ok(network.clone())
     }
 }
